requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.1
django-appconf==1.0.5
django-compressor==4.1
django-redis==5.2.0
djangorestframework==3.13.1
docker==6.0.0
idna==3.3
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.2.1
rcssmin==1.1.0
redis==4.3.4
requests==2.28.1
rjsmin==1.2.0
sqlparse==0.4.2
stripe==5.0.0
urllib3==1.26.11
websocket-client==1.3.3
whitenoise==6.2.0
wrapt==1.14.1
